requirements.txt
Django==3.2.12
django-globus-portal-framework==0.4.3
djangorestframework==3.13.1
humanize==4.0.0